Move retrieveWardVisitTotals to gateway
