+++ src/containers/AppContainer.js
@@ -89,6 +89,7 @@ import retrieveManagerByUuidGateway from "../gateways/MsSQL/retrieveManagerByUui
 import updateManagerStatusByUuid from "../usecases/updateManagerStatusByUuid";
 import updateManagerStatusByUuidGateway from "../gateways/MsSQL/updateManagerStatusByUuid";
 import archiveManagerByUuidGateway from "../gateways/MsSQL/archiveManagerByUuid";
+import retrieveWardVisitTotalsGateway from "../gateways/PostgreSQL/retrieveWardVisitTotals";
 
 class AppContainer {
   getDb = () => {
@@ -350,6 +351,10 @@ class AppContainer {
 
   getRetrieveHospitalsByTrustIdGateway = () => {
     return retrieveHospitalsByTrustIdGW(this);
+  };
+
+  getRetrieveWardVisitTotalsGateway = () => {
+    return retrieveWardVisitTotalsGateway(this);
   };
 
   getInsertHospitalGateway = () => {
